botbrew_commons/pyproject.toml
[tool.poetry]
name = "botbrew_commons"
version = "0.1.0"
description = "Shared tools"
authors = ["Leopoldo Garcia Vargas <lk13.dev@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
pydantic = "^2.5.3"
#GCP
google-cloud-secret-manager = "^2.18.0"
google_cloud_storage = "^2.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"